flesh out competition views a bit
add an individual competition page and fill out a basic index page, along
with the corresponding routes. Once we create a competition we want to
redirect to the competition that was just created. There's a lot more
functionality to add to the competition show page but for now we're content
with just displaying the information.